requirements.txt
gradio==4.19.2
groq==1.0.0
google-generativeai==0.8.6
python-dotenv==1.2.1
